+++ tests/epyccel/test_docstrings.py
@@ -60,3 +60,27 @@ def test_class_docstring(language):
 
     python_doc, pyccel_doc = pad_docstrings(A.__doc__, B.__doc__)
     assert python_doc == pyccel_doc
+
+def test_property_docstring(language):
+    class MyA:
+        """
+        Class containing x
+        """
+        def __init__(self : 'MyA', x : int):
+            self._x = x
+
+        @property
+        def x(self):
+            """
+            This is a property it cannot be set.
+            """
+            return self._x
+
+    B = epyccel(MyA, language=language)
+
+    print(MyA.__doc__, B.__doc__)
+
+    python_doc, pyccel_doc = pad_docstrings(MyA.__doc__, B.__doc__)
+    assert python_doc == pyccel_doc
+    python_doc, pyccel_doc = pad_docstrings(MyA.x.__doc__, B.x.__doc__)
+    assert python_doc == pyccel_doc